Use over-sampling to avoid branches in 2D blur effect
It refactors the bucketing and static variables into a template
function. Now that both 1D and 2D are defined in terms of an already
calculated kernel size, the bucketing on multiples of 4 and 8 is the
same.

Since the 2D effect packs 4 weights into a single half4, the 4-wise
bucketing fits fairly well. This reworks the effect to take another
half4 array of packed offsets that hold the 2D kernel offsets that
had been calculated based on the loop index and with branches.

Given the apparent cost of branches on these GPUs, I think this space
tradeoff will probably be worth it. To compensate for this extra
complex uniform, a new utility function is calculated to fill out the
2D array, carrying the last true offset to the over-sampled offsets
like I'd done in the 1D case previously

Bug: b/297590025
Change-Id: I42ac74280772492915c65cbc40aca31516045cbb
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/750343
Reviewed-by: Brian Osman <brianosman@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>