Require that texture2D function parameters include a pixel format.
We cannot pass a writable storage texture in SPIR-V or WGSL without
knowing its pixel format. (Metal does not allow reinterpreting the
pixel format of a texture; how you bound it is how it will be used.)

At present we do not enforce that passing a storage texture to a
helper function uses the correct pixel format; I've added an error
test for this which currently does not trigger errors. This is
addressed in a followup.

Bug: b/295160569
Bug: b/295157301
Change-Id: I6536f9e7c74dd1e2530ab294436e410e81eeac19
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/750402
Reviewed-by: Brian Osman <brianosman@google.com>
Commit-Queue: Brian Osman <brianosman@google.com>
Auto-Submit: John Stiles <johnstiles@google.com>