Enforce matching pixel-format when passing texture2D arguments.
This allows the frontend to detect mismatches between texture
pixel formats in a function-call argument list and the function's
actual parameter list. A `texture2D`-typed expression with a
different pixel format would be treated as a fundamentally
different type in most GPU APIs. This check needs to be
handled with care, because some modifier flags like `const` are
allowed to differ between argument and parameter.

http://go/samplers-and-storage-textures-in-sksl#heading=h.ia142csj1140

Bug: b/295160569
Bug: b/295157301
Change-Id: Ic9c12faf426b8eafb84fbc62dcf5127bed1dbb42
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/750592
Reviewed-by: Brian Osman <brianosman@google.com>
Commit-Queue: John Stiles <johnstiles@google.com>