Add support for pixel-format flags in SPIR-V backend.
The `getType` function now needs to know the Layout along with the
Type and MemoryLayout of the object. This required a lot of little
changes throughout the code generator, adding `const Layout&` to
various methods that previously only needed a type, and making sure
not to drop the layout as we pass types around.

Change-Id: I37e7861a838926b213f3dd3292fa1df380537a9c
Bug: b/293670098
Bug: b/295160569
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/751520
Auto-Submit: John Stiles <johnstiles@google.com>
Commit-Queue: Arman Uguray <armansito@google.com>
Reviewed-by: Arman Uguray <armansito@google.com>