Add epsilon to texture effect clamping at pixel centers
When sampling at pixel centers with bilinear filtering and clamping to
a subset, the last row or column could sample the adjacent exterior
row or column with a weight of 0. This extra inset ensures we sample
the previous row or column with a weight of 1 instead to avoid
including undefined data into a weighted sum (even if it shouldn't be
making a contribution).

Bug: chromium:1324336
Change-Id: I4e038c90e8c89ef535b4cc2f49c834960640a68a
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/751778
Reviewed-by: Greg Daniel <egdaniel@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>