[paragraph] Publicly depend on skia and skunicode
The public includes of skparagraph include files from skia an skunicode,
so these should be marked as public dependencies.

This also resolves issues with running ParagraphSlide1 when
SK_UNICODE_CLIENT_IMPLEMENTATION is defined. ParagraphBuilderImpl is a
different size if this is defined and it is currently defined while
building the skparagraph target because it directly depends on
skunicode, but it is not set when building the skparagraph slides
because they do not currently depend directly on skunicode.

Change-Id: Iec33ace6deef704755f9b18a7ef6d293822aea10
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/752617
Commit-Queue: Ben Wagner <bungeman@google.com>
Reviewed-by: Julia Lavrova <jlavrova@google.com>