Flatten clip virtuals in SkDevice
Since this is an internal type and the non-virtual public functions
weren't doing anything meaningful, I made them virtual and deleted the
protected virtuals. I also moved some of the semi-related bounds and
transform functions to be near the clip functions in the header. The
only one that remains is `onClipShader()` because the public one
handles mapping the SkClipOp. In this case, it was also moved to
private instead of protected.

I renamed `onClipAsRgn` to `android_utils_clipAsRgn` to emphasize its
limited use case.

I renamed `save()` and `restore()` to `pushClipStack()` and
`popClipStack()` to emphasize that it only affects the clip stack
(unlike in SkCanvas).

Bug: b/40042366
Bug: b/299474380
Change-Id: Id0e0efee4b1554a3dff8422706281d3de9877906
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/752459
Reviewed-by: Herb Derby <herb@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>