[Fontations] Add fetch script for retrieving Google Fonts test data
Fetch an archive of per-language test strings, and a mapping of
language coverage to font names and files including the font files
themselves. This archive is produced by an internal tool [1]
and consists of only open-source SIL Open Font License licensed
fonts, currently only from the Noto set.

Name a revision in DEPS and add a fetch-fonts-testdata script
for local development usage. In testing on infra, other means for
fetching the testdata CIPD archive are expected to be used, such
as a bazel-driven download.

Preparation for an upcoming CL which uses the test data to iterate
over a set of languages, retrieves the language-specific test
strings and performs path equality and rasterisation tests.

[1] cl/559229536

Bug: skia:295807418
Change-Id: I4fc002deec60c03dfe1ea708cc2656df0e584878
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/752837
Commit-Queue: Dominik Röttsches <drott@google.com>
Reviewed-by: Kevin Lubick <kjlubick@google.com>