Enable rgba syntax for SVG gradient stop colors
Currently, when providing an SVG object with a gradient that has a gradient-stop with rgba value an error occurs and it is rendered as black. This fixes it so that rgba colors can now be read and properly multiplied by the stop opacity. i.e. Having 0.5 alpha and 0.5 stop opacity should result in a 0.25 opacity for the final rendered color.

This is an imported pull request from
https://github.com/google/skia/pull/128

GitOrigin-RevId: 8b2f03788152587b1052cbce9d7e7d003d5cff3a
Change-Id: Ia5d24361314cd4a4410791f302bfa938a1d16ad9
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/657976
Commit-Queue: Florin Malita <fmalita@google.com>
Reviewed-by: Florin Malita <fmalita@google.com>