Implement sk_SampleMask and sk_SampleMaskIn in Metal.
`sk_SampleMaskIn` is a new input parameter to main, and
`sk_SampleMask` is now part of the Outputs structure. These
are only added to the program when they're used.

(We don't have any unit tests which use these variables yet,
but I've checked the functionality in KodeLife and they appear
to behave exactly as you'd hope.)

Bug: b/299484410
Change-Id: I0e62bcfd41b7b133aa07944ef4b3925680c43a36
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/753556
Auto-Submit: John Stiles <johnstiles@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>