[graphite][dawn] Use SkSL WGSL backend with compute pipelines
Enabled the SkSL->WGSL compilation when creating a compute pipeline
on Dawn.

To enable this, I had to remove the uses of "readonly" storage buffer
resource declarations in the test ComputeSteps. This is because
Graphite's suballocation of storage buffers does not play well with
WebGPU's buffer usage scopes which spans a whole buffer and leads
to validation errors when a read-only and a read-write buffer binding
for an entry-point is backed by the same buffer object.

Bug: b/40044665
Bug: b/40044196
Bug: b/299979165
Change-Id: I93ea25b1a28cf74c9db0cc0692d2a489e4c7d8f6
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/753846
Reviewed-by: John Stiles <johnstiles@google.com>
Commit-Queue: John Stiles <johnstiles@google.com>