Clamp alpha after doing CPU color filtering
I managed to trigger an assert when we applied the color filter to the
paint color. (My color filter happened to return colors with an alpha
greater than 1). The correct answer (and where to do the clamp) is
slightly vague, but I think this answer is closest to what we'd get if
we did a single-pixel draw plus a read-back.

Change-Id: I28b364c93b9ef47d8015eb4f64032a3ef02f4312
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/756056
Commit-Queue: Brian Osman <brianosman@google.com>
Reviewed-by: John Stiles <johnstiles@google.com>