[skif] Guard case when dst range starts and ends before src range
After https://skia-review.googlesource.com/c/skia/+/751776, dstIdx
could reach dstEnd and dstEnd might be < srcIdx. In that case srcIdx
and dstIdx aren't in sync, but we also have the case where dstIdx
was not less than the commonEnd calculated for the loop right after
the assert. So we were doing the right thing in release builds but
checking too soon on DCHECK builds.

Bug: b/1480853
Change-Id: I14ec662750259e8666aecc95e5be954754706ad2
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/756318
Reviewed-by: Robert Phillips <robertphillips@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>