Update SetTraceValue() to use a template, remove bitpun union
This reads a bit cleaner and allows for trace-using code to pass
types in more easily. The type union is removed in favor of static
casting or bit casting to avoid the can of worms that is C++ unions.

Change-Id: Ib0a19639e438d9552599ccd97c1854d25d344888
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/755944
Auto-Submit: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>