[graphite][compute] Coverage mask format based on storage binding support
The WebGPU standard disallows the R8Unorm format in storage bindings.
Compute mask renderers require the coverage mask texture to be bound
as a storage binding and this limitation forces us to use a RGBA8 type
texture on the Dawn backend.

Graphite currently uses the full compositing version of the vello fine
stage to rasterize coverage masks. This stage declares its target
texture as a `texture_storage_2d<rgba8unorm, write>` in the source WGSL.
When compiled to Metal, this target is always declared as a
`texture2d<float>` which happens to be compatible with the R8Unorm
format.

Given the above, we can continue to use the kAlpha_8_SkColorType on
Metal. This determination is now based on Caps: if Alpha8 is compatible
with storage then we pick that. Otherwise we default to RGBA8.

Bug: b/280932055
Change-Id: Ic5a1d7daed02a71820264d8bc73907d6fb52dcf3
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/757957
Reviewed-by: Jim Van Verth <jvanverth@google.com>
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: Arman Uguray <armansito@google.com>