[graphite] Ensure CoverageMaskShapes are created with a valid proxy
Previously a PathAtlas texture proxy was only created lazily when a
shape gets added to the atlas. Inverse fills with a completely clipped
out shape do not render a mask into the atlas but are still drawn as a
CoverageMaskShape, which needs to be constructed with a pointer to a
valid atlas texture proxy.

Now we make sure that a texture proxy is always available when a
CoverageMaskShape gets constructed. The texture proxy is still
initialized lazily rather than in the constructor, keeping in mind
that PathAtlas will eventually support multiple transient target
textures.

To test this change:
1. Run viewer -b grmtl --redraw --svgs test-svgs -m cubicpath_shader
2. Zoom out the view until one of the inverse fill shapes along the
   sides are partially outside the viewport, such that the shape is
   not visible except for an inverse fill region its bounding rect.

Change-Id: Ia8a8c0b730104e08b31502c6ba4daa9d1e53cac3
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/758117
Reviewed-by: Jim Van Verth <jvanverth@google.com>
Commit-Queue: Arman Uguray <armansito@google.com>