[graphite][dawn] Fix resource leak of GPU-only storage
DrawBufferManager was forgetting to add a resource ref to a buffer on
platforms that require a transfer buffer (such as Dawn) if the buffer
does not have an accompanying transfer buffer when the ownership of
buffers get passed on to the Recording.

This is possible for buffers with a GPU-only access mode, which never
need a CPU-to-GPU transfer buffer.

Change-Id: Ie2d4a2b051e2f9af633d1f4fbc7bf31484dd8456
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/759677
Commit-Queue: Arman Uguray <armansito@google.com>
Reviewed-by: Michael Ludwig <michaelludwig@google.com>