[skif] Draw FilterResults directly for layer restoration
Many of the upfront optimization analysis is handled automatically in
applyColorFilter() and drawing with the blender. This means that when
we have a restore that floods the clip we don't force the input image
to flood as well, we can apply that at the very end during rendering.

Additionally, any deferred effects like color filters and matrix
transforms from the final filters in the DAG can be applied when
drawing directly to the restore target device. Eventually when we get
to deferring the shader evaluation, this will further allow a filters
to avoid intermediate renderpasses.

The overall logic is simpler now as well, and the empty input filter
case can reuse internalDrawDeviceWithFilter() now too.

Bug: b/40040587
Bug: b/40042615
Change-Id: I12c452918f2229abdc55b4df4a18a1ccc2abf3aa
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/759117
Reviewed-by: Brian Osman <brianosman@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>