Expand //:core and introduce //:png_decode_codec et al
This adds hopefully all the pieces needed to use the "core" aspects
of Skia. In this sense, the "core" module refers to more than just
the .cpp files that are in //src/core and exposed from //include/core.

Specifically, we have decided the following things are "core":
 - include/core (e.g. SkPaint, SkCanvas, SkImage, SkFonts)
 - include/effects
 - include/sksl
 - include/utils (currently a subset, may grow)

To make the above things work, we must pull in files from
various subfolders of //src (see //src/core:core for more details),
including:
 - RasterPipeline (e.g. the CPU backend)
 - Raster-backed SkImages
 - Shaders (powered by sksl)

We could perhaps decouple core and RasterPipeline a bit more
in the future, making most of the CPU backend an optional add-in,
but that is beyond the scope of the project for now.

Follow-on CLs will:
 - Add CI jobs to test these modules (both using the Skia toolchain
   and a default toolchain)
 - Move test binaries to use the modules and run these on the CI.
 - Add more modules (e.g. more Font Manager options, GPU backends)

Suggested review order:
 - example/external_client/ to see the sample code and how it looks
   from the client POV
 - src/core/BUILD.bazel to see all the pieces
 - SkOpts related code to see some minor tweaks to how skopts is
   built (had to disconnect it from the legacy build's include rules)
 - All other BUILD.bazel files in any order.
 - bazel/ to see dependency-related minor changes.

Change-Id: I13eca3e4aa73016524533e653c1259c69063b381
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/753176
Reviewed-by: Ben Wagner <bungeman@google.com>
Commit-Queue: Kevin Lubick <kjlubick@google.com>