Add CI jobs to test external_client build of "buffet" options
I needed to add the explicit "-lpthread" link option due to GCC 10
(on the Debian docker image) being a bit older than GCC 13 on my
desktop (the previous "test" mechanism).

Rather than use the existing docker.py recipe, I made a new taskdriver
for this to reduce use of the recipe system.

I did have to update the Docker11 image due to needing a few more
dependencies when downloading bazelisk and using it.

I chose to put this in BazelTest because that already has a mechanism
for specifying a custom taskdriver.

Change-Id: I3662dd90cb125b5333ed798bbfaf5abbc7f5bfce
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/761376
Commit-Queue: Kevin Lubick <kjlubick@google.com>
Reviewed-by: Leandro Lovisolo <lovisolo@google.com>