[skif] Refactor Functors,ContextInfo,Context into virtual Backend, and Context
We used to have Context passed around by value, and it would hold a
ContextInfo, which was a mix of data that could change during filter
evaluation and data that didn't (e.g. cache and color type). All of
the backend specific functionality was then stored as function pointers
on a Functors struct.

The Functors struct stored everything as std::function and that
passed around by value. Besides lots of code size introduced in
generating std::function templates, this became cumbersome tweaking
the signatures or adding new capabilities to the backend adapter.

Now, Functors is replaced with a virtual Backend class that has the
same function capabilities that are then defined per raster, ganesh,
and graphite. The backend also now holds the unchanging data pulled
off of ContextInfo (color type, surface props, and the cache). The
image filter cache is also now created directly by the per-backend
creation functions instead of being a separate virtual on SkDevice.

Context now holds the rest of the data that was in ContextInfo, as
well as a pointer to the Backend. This breaks the weird instantiation
dependency loop that had been the case where you'd create the
Functors, create the ContextInfo with the empty source image, and
then use that to make a Context. Once you had a Context you could
make the real source image, and then create a second true Context.

Now you just create the backend, use it to set your state up, and
then create the Context.

Bug: b/302168680
Bug: b/40040586
Change-Id: Ibf8f0e7c11984b09782341b6c3f0757e97fc44a9
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/760740
Reviewed-by: Robert Phillips <robertphillips@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>