Incorporate child effects into the mesh key and proxy handlers.
Including child effects in the key is necessary to ensure that we
do not reuse a mismatched program from the cache. The proxy handler
is required to properly set up dependencies on the referenced
surfaces--the Vulkan backend asserts if you do not do this, but
OpenGL and Metal appear to silently allow it.

Bug: b/40045302
Change-Id: Ia6cf5ca6f0adb042318e76a1f03aa90064e29814
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/762996
Reviewed-by: Jim Van Verth <jvanverth@google.com>
Auto-Submit: John Stiles <johnstiles@google.com>
Commit-Queue: John Stiles <johnstiles@google.com>