Remove GMVerifier class from DM.
This feature was landed in early 2020 as a way for GM images to
report failures if they mismatched a known source of truth. However,
in three years, this feature has not been used anywhere, and its
author is no longer on the team. Let's clean it up.

Bug: b/40041184
Change-Id: Ided8013e53cecd61ddb987f279cb387d0bc6dbc5
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/764519
Commit-Queue: John Stiles <johnstiles@google.com>
Reviewed-by: Derek Sollenberger <djsollen@google.com>
Auto-Submit: John Stiles <johnstiles@google.com>