[skif] Use content bounds for leaf input layer sizing
I believe this should fix the numerous issues with the DDL GM configs
and image filters. When the content bounds would be defined beyond the
desired output (which was frequently the case with how the DDL sink
would clip and tile the scene), the image filter should have been
requiring that larger content bounds at the leaf level instead of just
requesting the desired output and restricting it after the fact to
a possibly-smaller content bounds.

Bug: b/40040615
Change-Id: I2980d0ff9a0975b1a3ba2f5e7caffb5270d27a65
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/744008
Reviewed-by: Robert Phillips <robertphillips@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>