Fix inlining of calls with unused inout params.
Previously, we would fail to create a scratch variable because of
the unexpected `out` modifier flag. In practice, it's safe to
ignore this, and it's necessary to create a scratch variable
to preserve the argument if it has side-effects that must be
preserved. (In the final output, the variable is actually optimized
away, because it was never read from or written to, but the side
effect is preserved, and that is the important part.)

Bug: oss-fuzz:63170
Change-Id: I953e45eb28545ff94e20e91e6c4eb699c0a9c3bf
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/765397
Reviewed-by: Arman Uguray <armansito@google.com>
Auto-Submit: John Stiles <johnstiles@google.com>
Commit-Queue: John Stiles <johnstiles@google.com>