[bazel] Update version of rules_pkg
The dependency had not been explicitly declared, so it was indirectly
installed via a different dep and was very old (0.2.1).

The old version was shelling out to Python2 (!) to package up the
tar used in the images. This new version uses a Python3 script.

Python2 stopped working when we updated to Bazel 6 [1], probably because
a Python2 toolchain was not implicitly downloaded and registered.
This caused our Louhi flows to stop working, causing the autobuilds
(e.g. skottie.skia.org) to not be updated.

[1] https://skia-review.googlesource.com/c/skia/+/729416

No-Try: true
Change-Id: I7fc4bfa44f0a1be62c3969046cfad1da2dd19650
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/765758
Reviewed-by: Leandro Lovisolo <lovisolo@google.com>
Auto-Submit: Kevin Lubick <kjlubick@google.com>
Commit-Queue: Leandro Lovisolo <lovisolo@google.com>