graphite: Fix SkMipmap memory leak
MakeBitmapProxyView calls SkMipmap::Build which returns a SkMipmap* with
a ref count of 1. Calling sk_ref_sp on it will increment its ref count
to 2 and the initial ref count will never be decremented and the memory
will be leaked. Fix this by wrapping the SkMipmap in a sk_sp ("adopt"
semantics) instead of using sk_ref_sp in MakeBitmapProxyView.

Bug: b/238754357
Change-Id: Id76d41528fdbf6ff1f1ad6b753c53b0971e56ef3
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/766076
Reviewed-by: Greg Daniel <egdaniel@google.com>
Commit-Queue: Sunny Sachanandani <sunnyps@google.com>