Fix fuzzer-discovered assertion with atomic arrays.
We did not properly reject a vardecl which attempted to set an
initializer-expression on an atomicUint array. However, we would
reject the equivalent BinaryExpression assignment. This led to an
assertion in BinaryExpression::Make when rewriting the vardecl as
an assignment in Transform::HoistSwitchVarDeclarationsAtTopLevel.

Bug: oss-fuzz:63411
Change-Id: Iac0b57bb227839ff8d1a0ef5b4d0b4ec28117626
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/769092
Commit-Queue: John Stiles <johnstiles@google.com>
Commit-Queue: Brian Osman <brianosman@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>
Auto-Submit: John Stiles <johnstiles@google.com>