[graphite] Write depth*w for text and cover bounds VS functions
When there's perspective, the VS needs to write depth*W so that the
value post homogeneous division is a constant across the geometry of
draw (since the clipping/painters-order system expects that regardless
of perspective).

This ends up causing some text pixel diffs under perspective because
of the text RenderSteps' use of GEQUAL. There were different depth-test
pixel issues previously because glyphs in the same subrun that
overlapped would have non-constant depth functions and then it was a
bit unpredictable as to which glyph would win each pixel.

Now, each glyph writes a constant depth, but every glyph has the same
paint depth. Under perspective, and a 32-bit float depth buffer, there
is actually *too* much precision and some glyphs that come later in
the subrun have an epsilon-smaller depth and fail the depth test.
The second listed bug captures plans to address this z-fighting
amongst planar glyphs that overlap. However, this CL still leads to
a more correct (but still not right) behavior so it's worth landing
on its own.

Bug: b/266204364
Bug: b/306501391
Change-Id: I0e6c6c52eadce30115ea044107aff90858ad1494
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/769897
Reviewed-by: Arman Uguray <armansito@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>