[graphite] Handle atlas padding in Rectanizer, simpify PathAtlas API
The CoverageMaskShape's MaskInfo now has UVs and sizes that define the
regular mask size, not the padded boundary. This simplifies the
arithmetic in the vertex shader and lets the RenderStep be reusable
for masks that may not have had padding to begin with.

The base PathAtlas does not handle any of the padding calculations,
and for the most part, neither do the subclasses. Instead Rectanizer
now has a helper `addPaddedRect` that increases the alloted size
by 2*padding, and offsets the assigned position by padding.

The rest of the transform math used to render shapes is then simplified
to translate the logical (0,0,w,h) draw by the rectanizer-assigned
position.

Change-Id: I70f687b98eae48fd0b9bfd1cfdb81ac1d6f737a7
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/768296
Commit-Queue: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Jim Van Verth <jvanverth@google.com>