[graphite] Generalize transform handling in CoverageMaskRenderStep
Device::drawCoverageMask will need to draw the mask with an arbitrary
final mask-to-device transform, but the uniform and instance attribute
structure is set up to optimize batching for PathAtlas-produced
CoverageMaskShapes since extracting a deviceOrigin translation
leaves the uniform matrix as the identity.

With the addition of a full matrix to the vertex shader, I opted to
normalize the mask coordinates on the CPU and then apply a
de-normalizing scale factor to the mask-to-device matrix. To make
sure the same amount of instance attribute space was used as before,
I encoded the mask bounds as ushort4_norm. Other attributes that had
been full ints I moved to float since they were always just being
cast to float in the vertex shader anyways.

The other meaningful change to the shader was that I removed the
inverted-ness uniform. The mask bounds now swap their sorting for
inverse fills and then a half varying is used to mix c and (1-c)
for a branch-less fragment shader. We probably would have been seeing
decent branch prediction on a uniform before, but now we can combine
inverse and non-inverse-fill draws.

Change-Id: Iebb96f87a90b1c34eab76f04640161725fdb17fe
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/768356
Reviewed-by: Arman Uguray <armansito@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>