Add position storage to SkColor4fXformer
At this stage, we just use this for the "make positions explicit"
request from Ganesh. However, upcoming changes for polar interpolation
spaces will require injecting new hard-stops, so the xformer will have
to have a mutable copy of the positions. This is also why I adjusted all
the users to refer to the xformer to get the color-count and positions.

Bug: chromium:1462612
Change-Id: I29dd601b4373618ede0dc0cb85cf60ec6fddabe4
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/771478
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: Brian Osman <brianosman@google.com>