Don't apply pragma pack to SkOTTableName::Iterator
pragma pack is hard to use correctly. Because the individual fields in
the struct are incorrectly aligned, you cannot access any of those
fields by pointer or reference. The pointer or reference will be
unaligned, and won't be captured in the type system.

SkOTTableName has pragma pack applied for the benefit of the OpenType
table structures. But it later also grew a C++ iterator type, containing
SkString. Those types cannot be packed, or we hit this issue.

It's a little awkward because Skia made this an inner type, but we can
fix this by turning the pragma off a little earlier.

Bug: chromium:1394755
Change-Id: I33520f63db8a879537da647c16df5a62c309a666
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/771700
Reviewed-by: Ben Wagner <bungeman@google.com>
Commit-Queue: Ben Wagner <bungeman@google.com>
Auto-Submit: David Benjamin <davidben@google.com>